Fix Effects Lab selection bug
