+++ PowerPointLabs/PowerPointLabs/EffectsLab/EffectsLabBlur.cs
@@ -2,12 +2,10 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
-using System.Windows.Forms;
 
 using ImageProcessor;
 using ImageProcessor.Imaging;
 using PowerPointLabs.CropLab;
-using PowerPointLabs.Views;
 
 using Office = Microsoft.Office.Core;
 using PowerPoint = Microsoft.Office.Interop.PowerPoint;
@@ -24,14 +22,8 @@ namespace PowerPointLabs.EffectsLab
         public static int CustomPercentageRemainder = 30;
         public static int CustomPercentageBackground = 30;
 
-        private static Models.PowerPointSlide _slide;
-
         private const string HexColor = "#000000";
         private const float Transparency = 0.8f;
-
-        private const string MessageBoxTitle = "Error";
-        private const string ErrorMessageNoSelection = TextCollection.EffectsLabBlurSelectedErrorNoSelection;
-        private const string ErrorMessageNonShapeOrTextBox = TextCollection.EffectsLabBlurSelectedErrorNonShapeOrTextBox;
 
         private static readonly string BlurPicture = Path.GetTempPath() + @"\blur.png";
 
@@ -60,15 +52,13 @@ namespace PowerPointLabs.EffectsLab
 
             try
             {
-                _slide = slide;
-
                 var hasManyShapes = shapeRange.Count > 1;
                 var shape = hasManyShapes ? shapeRange.Group() : shapeRange[1];
                 var left = shape.Left;
                 var top = shape.Top;
                 shapeRange.Cut();
 
-                Utils.Graphics.ExportSlide(_slide, BlurPicture);
+                Utils.Graphics.ExportSlide(slide, BlurPicture);
                 BlurImage(BlurPicture, percentage);
 
                 shapeRange = slide.Shapes.Paste();
@@ -79,9 +69,9 @@ namespace PowerPointLabs.EffectsLab
                     shapeRange = shapeRange.Ungroup();
                 }
 
-                var ungroupedRange = UngroupAllShapeRange(shapeRange);
-                var shapeGroupNames = ApplyBlurEffect(BlurPicture, ungroupedRange);
-                var range = _slide.Shapes.Range(shapeGroupNames.ToArray());
+                var ungroupedRange = EffectsLabUtil.UngroupAllShapeRange(slide, shapeRange);
+                var shapeGroupNames = ApplyBlurEffect(slide, BlurPicture, ungroupedRange);
+                var range = slide.Shapes.Range(shapeGroupNames.ToArray());
 
                 return range;
             }
@@ -89,14 +79,14 @@ namespace PowerPointLabs.EffectsLab
             {
                 ActionFramework.Common.Log.Logger.LogException(e, "BlurSelectedEffect");
 
-                ShowErrorMessageBox(e.Message, e);
+                EffectsLabUtil.ShowErrorMessageBox(e.Message, e);
                 return null;
             }
         }
 
         public static void BlurRemainder(Models.PowerPointSlide slide, PowerPoint.Selection selection, int percentage)
         {
-            var effectSlide = EffectsLabRecolor.GenerateEffectSlide(slide, selection, true);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(slide, selection, true);
 
             if (effectSlide == null)
             {
@@ -109,7 +99,7 @@ namespace PowerPointLabs.EffectsLab
 
         public static void BlurBackground(Models.PowerPointSlide slide, PowerPoint.Selection selection, int percentage)
         {
-            var effectSlide = EffectsLabRecolor.GenerateEffectSlide(slide, selection, false);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(slide, selection, false);
 
             if (effectSlide == null)
             {
@@ -159,7 +149,7 @@ namespace PowerPointLabs.EffectsLab
             }
             else
             {
-                overlayShape = DuplicateShapeInPlace(blurShape);
+                overlayShape = EffectsLabUtil.DuplicateShapeInPlace(blurShape);
             }
 
             Utils.Graphics.MoveZToJustInFront(overlayShape, blurShape);
@@ -184,7 +174,7 @@ namespace PowerPointLabs.EffectsLab
                 return true;
             }
 
-            ShowErrorMessageBox(ErrorMessageNoSelection);
+            EffectsLabUtil.ShowNoSelectionErrorMessage();
             return false;
         }
 
@@ -195,51 +185,11 @@ namespace PowerPointLabs.EffectsLab
                 return true;
             }
 
-            ShowErrorMessageBox(ErrorMessageNoSelection);
+            EffectsLabUtil.ShowNoSelectionErrorMessage();
             return false;
         }
 
-        private static PowerPoint.ShapeRange UngroupAllShapeRange(PowerPoint.ShapeRange shapeRange)
-        {
-            var ungroupedShapeNames = new List<string>();
-            var queue = new Queue<PowerPoint.Shape>();
-
-            foreach (PowerPoint.Shape shape in shapeRange)
-            {
-                queue.Enqueue(shape);
-            }
-
-            while (queue.Count != 0)
-            {
-                var shape = queue.Dequeue();
-
-                if (shape.Type == Office.MsoShapeType.msoGroup)
-                {
-                    var subRange = shape.Ungroup();
-                    foreach (PowerPoint.Shape item in subRange)
-                    {
-                        queue.Enqueue(item);
-                    }
-                }
-                else if (shape.Type == Office.MsoShapeType.msoPlaceholder
-                    || shape.Type == Office.MsoShapeType.msoTextBox
-                    || shape.Type == Office.MsoShapeType.msoAutoShape
-                    || shape.Type == Office.MsoShapeType.msoFreeform)
-                {
-                    ungroupedShapeNames.Add(shape.Name);
-                }
-                else
-                {
-                    throw new Exception(ErrorMessageNonShapeOrTextBox);
-                }
-            }
-
-            var ungroupedShapeRange = _slide.Shapes.Range(ungroupedShapeNames.ToArray());
-
-            return ungroupedShapeRange;
-        }
-
-        private static List<string> ApplyBlurEffect(string imageFile, PowerPoint.ShapeRange shapeRange)
+        private static List<string> ApplyBlurEffect(Models.PowerPointSlide slide, string imageFile, PowerPoint.ShapeRange shapeRange)
         {
             var shapeGroupNames = new List<string>();
 
@@ -250,12 +200,12 @@ namespace PowerPointLabs.EffectsLab
                 if (shape.Type == Office.MsoShapeType.msoPlaceholder
                     || shape.Type == Office.MsoShapeType.msoTextBox)
                 {
-                    var shapeNames = ApplyBlurEffectTextBox(imageFile, shape);
+                    var shapeNames = ApplyBlurEffectTextBox(slide, imageFile, shape);
                     shapeGroupNames.AddRange(shapeNames);
                 }
                 else // if (shape.Type == Office.MsoShapeType.msoAutoShape || shape.Type == Office.MsoShapeType.msoFreeform)
                 {
-                    var shapeName = ApplyBlurEffectShape(imageFile, shape);
+                    var shapeName = ApplyBlurEffectShape(slide, imageFile, shape);
                     shapeGroupNames.Add(shapeName);
                 }
             }
@@ -263,7 +213,7 @@ namespace PowerPointLabs.EffectsLab
             return shapeGroupNames;
         }
 
-        private static List<string> ApplyBlurEffectTextBox(string imageFile, PowerPoint.Shape textBox)
+        private static List<string> ApplyBlurEffectTextBox(Models.PowerPointSlide slide, string imageFile, PowerPoint.Shape textBox)
         {
             var shapeNames = new List<string>();
             shapeNames.Add(textBox.Name);
@@ -272,23 +222,23 @@ namespace PowerPointLabs.EffectsLab
             textBox.Fill.Visible = Office.MsoTriState.msoFalse;
             textBox.Line.Visible = Office.MsoTriState.msoFalse;
 
-            var blurShape = _slide.Shapes.AddShape(Office.MsoAutoShapeType.msoShapeRectangle, textBox.Left, textBox.Top, textBox.Width,
+            var blurShape = slide.Shapes.AddShape(Office.MsoAutoShapeType.msoShapeRectangle, textBox.Left, textBox.Top, textBox.Width,
                         textBox.Height);
             blurShape.Rotation = textBox.Rotation;
             Utils.Graphics.MoveZToJustBehind(blurShape, textBox);
-            CropToShape.FillInShapeWithImage(_slide, imageFile, blurShape, isInPlace: true);
+            CropToShape.FillInShapeWithImage(slide, imageFile, blurShape, isInPlace: true);
             shapeNames.Add(blurShape.Name);
             
             if (IsTintSelected)
             {
-                var overlayShape = GenerateOverlayShape(_slide, blurShape);
+                var overlayShape = GenerateOverlayShape(slide, blurShape);
                 shapeNames.Add(overlayShape.Name);
             }
 
             // cannot group placeholders
             if (textBox.Type != Office.MsoShapeType.msoPlaceholder)
             {
-                var subRange = _slide.Shapes.Range(shapeNames.ToArray());
+                var subRange = slide.Shapes.Range(shapeNames.ToArray());
                 var groupedShape = subRange.Group();
                 shapeNames.Clear();
                 shapeNames.Add(groupedShape.Name);
@@ -297,7 +247,7 @@ namespace PowerPointLabs.EffectsLab
             return shapeNames;
         }
 
-        private static string ApplyBlurEffectShape(string imageFile, PowerPoint.Shape shape)
+        private static string ApplyBlurEffectShape(Models.PowerPointSlide slide, string imageFile, PowerPoint.Shape shape)
         {
             var shapeNames = new List<string>();
             shapeNames.Add(shape.Name);
@@ -306,7 +256,7 @@ namespace PowerPointLabs.EffectsLab
             {
                 shape.ZOrder(Office.MsoZOrderCmd.msoBringToFront);
 
-                var textBox = DuplicateShapeInPlace(shape);
+                var textBox = EffectsLabUtil.DuplicateShapeInPlace(shape);
                 textBox.Fill.Visible = Office.MsoTriState.msoFalse;
                 textBox.Line.Visible = Office.MsoTriState.msoFalse;
                 Utils.Graphics.MoveZToJustInFront(textBox, shape);
@@ -314,52 +264,23 @@ namespace PowerPointLabs.EffectsLab
             }
 
             shape.TextFrame2.DeleteText();
-            CropToShape.FillInShapeWithImage(_slide, imageFile, shape, isInPlace: true);
+            CropToShape.FillInShapeWithImage(slide, imageFile, shape, isInPlace: true);
 
             if (IsTintSelected)
             {
-                var overlayShape = GenerateOverlayShape(_slide, shape);
+                var overlayShape = GenerateOverlayShape(slide, shape);
                 shapeNames.Add(overlayShape.Name);
             }
 
             if (shapeNames.Count > 1)
             {
-                var subRange = _slide.Shapes.Range(shapeNames.ToArray());
+                var subRange = slide.Shapes.Range(shapeNames.ToArray());
                 var groupedShape = subRange.Group();
 
                 return groupedShape.Name;
             }
 
             return shapeNames[0];
-        }
-
-        private static PowerPoint.Shape DuplicateShapeInPlace(PowerPoint.Shape shape)
-        {
-            var duplicateShape = shape.Duplicate()[1];
-            duplicateShape.Left = shape.Left;
-            duplicateShape.Top = shape.Top;
-
-            var match = System.Text.RegularExpressions.Regex.Match(duplicateShape.Name, @"\d+$");
-            if (!match.Success || int.Parse(match.Value) != duplicateShape.Id - 1)
-            {
-                duplicateShape.Name += " " + (duplicateShape.Id - 1);
-            }
-
-            return duplicateShape;
-        }
-
-        private static void ShowErrorMessageBox(string content, Exception exception = null)
-        {
-            if (exception == null
-                || content.Equals(ErrorMessageNoSelection)
-                || content.Equals(ErrorMessageNonShapeOrTextBox))
-            {
-                MessageBox.Show(content, MessageBoxTitle);
-            }
-            else
-            {
-                ErrorDialogBox.ShowDialog(MessageBoxTitle, content, exception);
-            }
         }
     }
 }
