+++ PowerPointLabs/PowerPointLabs/EffectsLab/EffectsLabRecolor.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Runtime.InteropServices;
-using System.Windows.Forms;
-
-using Microsoft.Office.Interop.PowerPoint;
+﻿using Microsoft.Office.Interop.PowerPoint;
 using PowerPointLabs.Models;
-using PowerPointLabs.Views;
 
 namespace PowerPointLabs.EffectsLab
 {
@@ -12,7 +7,7 @@ namespace PowerPointLabs.EffectsLab
     {
         public static void GreyScaleRemainderEffect(PowerPointSlide curSlide, Selection selection)
         {
-            var effectSlide = GenerateEffectSlide(curSlide, selection, true);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(curSlide, selection, true);
 
             if (effectSlide == null)
             {
@@ -25,7 +20,7 @@ namespace PowerPointLabs.EffectsLab
 
         public static void BlackWhiteRemainderEffect(PowerPointSlide curSlide, Selection selection)
         {
-            var effectSlide = GenerateEffectSlide(curSlide, selection, true);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(curSlide, selection, true);
             
             if (effectSlide == null)
             {
@@ -39,7 +34,7 @@ namespace PowerPointLabs.EffectsLab
         public static void GothamRemainderEffect(PowerPointSlide curSlide, Selection selection)
         {
 
-            var effectSlide = GenerateEffectSlide(curSlide, selection, true);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(curSlide, selection, true);
 
             if (effectSlide == null)
             {
@@ -52,7 +47,7 @@ namespace PowerPointLabs.EffectsLab
 
         public static void SepiaRemainderEffect(PowerPointSlide curSlide, Selection selection)
         {
-            var effectSlide = GenerateEffectSlide(curSlide, selection, true);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(curSlide, selection, true);
 
             if (effectSlide == null)
             {
@@ -65,7 +60,7 @@ namespace PowerPointLabs.EffectsLab
 
         public static void GreyScaleBackgroundEffect(PowerPointSlide curSlide, Selection selection)
         {
-            var effectSlide = GenerateEffectSlide(curSlide, selection, false);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(curSlide, selection, false);
 
             if (effectSlide == null)
             {
@@ -78,7 +73,7 @@ namespace PowerPointLabs.EffectsLab
 
         public static void BlackWhiteBackgroundEffect(PowerPointSlide curSlide, Selection selection)
         {
-            var effectSlide = GenerateEffectSlide(curSlide, selection, false);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(curSlide, selection, false);
 
             if (effectSlide == null)
             {
@@ -91,7 +86,7 @@ namespace PowerPointLabs.EffectsLab
 
         public static void GothamBackgroundEffect(PowerPointSlide curSlide, Selection selection)
         {
-            var effectSlide = GenerateEffectSlide(curSlide, selection, false);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(curSlide, selection, false);
 
             if (effectSlide == null)
             {
@@ -104,7 +99,7 @@ namespace PowerPointLabs.EffectsLab
 
         public static void SepiaBackgroundEffect(PowerPointSlide curSlide, Selection selection)
         {
-            var effectSlide = GenerateEffectSlide(curSlide, selection, false);
+            var effectSlide = EffectsLabUtil.GenerateEffectSlide(curSlide, selection, false);
 
             if (effectSlide == null)
             {
@@ -113,65 +108,6 @@ namespace PowerPointLabs.EffectsLab
 
             effectSlide.SepiaBackground();
             effectSlide.GetNativeSlide().Select();
-        }
-
-        internal static PowerPointBgEffectSlide GenerateEffectSlide(PowerPointSlide curSlide, Selection selection, bool generateOnRemainder)
-        {
-            PowerPointSlide dupSlide = null;
-
-            try
-            {
-                var shapeRange = selection.ShapeRange;
-
-                if (shapeRange.Count != 0)
-                {
-                    dupSlide = curSlide.Duplicate();
-                }
-
-                shapeRange.Cut();
-
-                var effectSlide = PowerPointBgEffectSlide.BgEffectFactory(curSlide.GetNativeSlide(), generateOnRemainder);
-
-                if (dupSlide != null)
-                {
-                    if (generateOnRemainder)
-                    {
-                        dupSlide.Delete();
-                    }
-                    else
-                    {
-                        dupSlide.MoveTo(curSlide.Index);
-                        curSlide.Delete();
-                    }
-                }
-
-                return effectSlide;
-            }
-            catch (InvalidOperationException e)
-            {
-                MessageBox.Show(e.Message);
-                return null;
-            }
-            catch (COMException)
-            {
-                if (dupSlide != null)
-                {
-                    dupSlide.Delete();
-                }
-
-                MessageBox.Show("Please select at least 1 shape");
-                return null;
-            }
-            catch (Exception e)
-            {
-                if (dupSlide != null)
-                {
-                    dupSlide.Delete();
-                }
-
-                ErrorDialogBox.ShowDialog("Error", e.Message, e);
-                return null;
-            }
         }
     }
 }
